+++ ts/features/messages/saga/__test__/handleLoadPreviousPageMessages.test.ts
@@ -3,24 +3,23 @@ import { testSaga } from "redux-saga-test-plan";
 import { getType } from "typesafe-actions";
 
 import {
-  loadNextPageMessages as action,
-  loadNextPageMessages
+  loadPreviousPageMessages as action,
+  loadPreviousPageMessages
 } from "../../store/actions";
 import {
   apiPayload,
   defaultRequestPayload,
-  successLoadNextPageMessagesPayload
+  successLoadPreviousPageMessagesPayload
 } from "../../__mocks__/messages";
-import { testTryLoadNextPageMessages } from "../watchLoadNextPageMessages";
 import { withRefreshApiCall } from "../../../fastLogin/saga/utils";
+import { handleLoadPreviousPageMessages } from "../handleLoadPreviousPageMessages";
+import { BackendClient } from "../../../../api/__mocks__/backend";
 
-const tryLoadNextPageMessages = testTryLoadNextPageMessages!;
-
-describe("tryLoadNextPageMessages", () => {
+describe("handleLoadPreviousPageMessages", () => {
   const getMessagesPayload = {
     enrich_result_data: true,
-    page_size: defaultRequestPayload.pageSize,
-    maximum_id: undefined,
+    page_size: 8,
+    minimum_id: undefined,
     archived: defaultRequestPayload.filter.getArchived
   };
 
@@ -28,19 +27,19 @@ describe("tryLoadNextPageMessages", () => {
     it(`should put ${getType(
       action.success
     )} with the parsed messages and pagination data`, () => {
-      const getMessages = jest.fn();
       testSaga(
-        tryLoadNextPageMessages(getMessages),
+        handleLoadPreviousPageMessages,
+        BackendClient.getMessages,
         action.request(defaultRequestPayload)
       )
         .next()
         .call(
           withRefreshApiCall,
-          getMessages(getMessagesPayload),
-          loadNextPageMessages.request(defaultRequestPayload)
+          BackendClient.getMessages(getMessagesPayload),
+          loadPreviousPageMessages.request(defaultRequestPayload)
         )
         .next(E.right({ status: 200, value: apiPayload }))
-        .put(action.success(successLoadNextPageMessagesPayload))
+        .put(action.success(successLoadPreviousPageMessagesPayload))
         .next()
         .isDone();
     });
@@ -48,16 +47,16 @@ describe("tryLoadNextPageMessages", () => {
 
   describe("when the response is an Error", () => {
     it(`should put ${getType(action.failure)} with the error message`, () => {
-      const getMessages = jest.fn();
       testSaga(
-        tryLoadNextPageMessages(getMessages),
+        handleLoadPreviousPageMessages,
+        BackendClient.getMessages,
         action.request(defaultRequestPayload)
       )
         .next()
         .call(
           withRefreshApiCall,
-          getMessages(getMessagesPayload),
-          loadNextPageMessages.request(defaultRequestPayload)
+          BackendClient.getMessages(getMessagesPayload),
+          loadPreviousPageMessages.request(defaultRequestPayload)
         )
         .next(E.right({ status: 500, value: { title: "Backend error" } }))
         .put(
@@ -73,14 +72,13 @@ describe("tryLoadNextPageMessages", () => {
 
   describe("when the handler throws", () => {
     it(`should catch it and put ${getType(action.failure)}`, () => {
-      const getMessages = () => {
-        throw new Error("I made a boo-boo, sir!");
-      };
       testSaga(
-        tryLoadNextPageMessages(getMessages),
+        handleLoadPreviousPageMessages,
+        BackendClient.getMessages,
         action.request(defaultRequestPayload)
       )
         .next()
+        .throw(new Error("I made a boo-boo, sir!"))
         .put(
           action.failure({
             error: new Error("I made a boo-boo, sir!"),
