+++ ts/features/messages/saga/__test__/handleUpsertMessageStatusAttribues.test.ts
@@ -1,19 +1,17 @@
 import * as E from "fp-ts/lib/Either";
 import { testSaga } from "redux-saga-test-plan";
 import { getType } from "typesafe-actions";
-
 import {
   upsertMessageStatusAttributes as action,
   UpsertMessageStatusAttributesPayload
 } from "../../store/actions";
 import { UIMessageId } from "../../types";
 import { successReloadMessagesPayload } from "../../__mocks__/messages";
-import { testTryUpsertMessageStatusAttributes } from "../watchUpsertMessageStatusAttribues";
 import { withRefreshApiCall } from "../../../fastLogin/saga/utils";
+import { handleUpsertMessageStatusAttribues } from "../handleUpsertMessageStatusAttribues";
+import { BackendClient } from "../../../../api/__mocks__/backend";
 
-const tryUpsertMessageStatusAttributes = testTryUpsertMessageStatusAttributes!;
-
-describe("tryUpsertMessageStatusAttributes", () => {
+describe("handleUpsertMessageStatusAttribues", () => {
   const actionPayload: UpsertMessageStatusAttributesPayload = {
     message: {
       ...successReloadMessagesPayload.messages[0],
@@ -35,15 +33,15 @@ describe("tryUpsertMessageStatusAttributes", () => {
     it(`should put ${getType(
       action.success
     )} with the original payload`, () => {
-      const putMessage = jest.fn();
       testSaga(
-        tryUpsertMessageStatusAttributes(putMessage),
+        handleUpsertMessageStatusAttribues,
+        BackendClient.upsertMessageStatusAttributes,
         action.request(actionPayload)
       )
         .next()
         .call(
           withRefreshApiCall,
-          putMessage(callPayload),
+          BackendClient.upsertMessageStatusAttributes(callPayload),
           action.request(actionPayload)
         )
         .next(E.right({ status: 200, value: {} }))
@@ -55,15 +53,15 @@ describe("tryUpsertMessageStatusAttributes", () => {
 
   describe("when the response is an Error", () => {
     it(`should put ${getType(action.failure)} with the error message`, () => {
-      const putMessage = jest.fn();
       testSaga(
-        tryUpsertMessageStatusAttributes(putMessage),
+        handleUpsertMessageStatusAttribues,
+        BackendClient.upsertMessageStatusAttributes,
         action.request(actionPayload)
       )
         .next()
         .call(
           withRefreshApiCall,
-          putMessage(callPayload),
+          BackendClient.upsertMessageStatusAttributes(callPayload),
           action.request(actionPayload)
         )
         .next(
@@ -85,14 +83,13 @@ describe("tryUpsertMessageStatusAttributes", () => {
 
   describe("when the handler throws", () => {
     it(`should catch it and put ${getType(action.failure)}`, () => {
-      const putMessage = () => {
-        throw new Error("462e5dffdb46435995d545999bed6b11");
-      };
       testSaga(
-        tryUpsertMessageStatusAttributes(putMessage),
+        handleUpsertMessageStatusAttribues,
+        BackendClient.upsertMessageStatusAttributes,
         action.request(actionPayload)
       )
         .next()
+        .throw(new Error("462e5dffdb46435995d545999bed6b11"))
         .put(
           action.failure({
             error: new Error("462e5dffdb46435995d545999bed6b11"),
