+++ ts/features/messages/saga/index.ts
@@ -1,35 +1,118 @@
 import { SagaIterator } from "redux-saga";
 import {
-  call,
+  fork,
   put,
   select,
   takeEvery,
   takeLatest
 } from "typed-redux-saga/macro";
-import { ActionType, getType } from "typesafe-actions";
+import { getType } from "typesafe-actions";
 import { SessionToken } from "../../../types/SessionToken";
 import { clearCache } from "../../../store/actions/profile";
 import { logoutSuccess } from "../../../store/actions/authentication";
 import {
   downloadAttachment,
   getMessageDataAction,
-  removeCachedAttachment
+  getMessagePrecondition,
+  loadMessageById,
+  loadMessageDetails,
+  loadNextPageMessages,
+  loadPreviousPageMessages,
+  loadThirdPartyMessage,
+  migrateToPaginatedMessages,
+  reloadAllMessages,
+  removeCachedAttachment,
+  upsertMessageStatusAttributes
 } from "../store/actions";
 import { retryDataAfterFastLoginSessionExpirationSelector } from "../store/reducers/messageGetStatus";
+import { BackendClient } from "../../../api/backend";
 import { handleDownloadAttachment } from "./handleDownloadAttachment";
 import {
   handleClearAllAttachments,
   handleClearAttachment
 } from "./handleClearAttachments";
 import { handleLoadMessageData } from "./handleLoadMessageData";
+import { handleLoadNextPageMessages } from "./handleLoadNextPageMessages";
+import { handleLoadPreviousPageMessages } from "./handleLoadPreviousPageMessages";
+import { handleReloadAllMessages } from "./handleReloadAllMessages";
+import { handleLoadMessageById } from "./handleLoadMessageById";
+import { handleLoadMessageDetails } from "./handleLoadMessageDetails";
+import { handleUpsertMessageStatusAttribues } from "./handleUpsertMessageStatusAttribues";
+import { handleMigrateToPagination } from "./handleMigrateToPagination";
+import { handleMessagePrecondition } from "./handleMessagePrecondition";
+import { handleThirdPartyMessage } from "./handleThirdPartyMessage";
 
-/**
- * Handle the message attachment requests
- * @param bearerToken
- */
-export function* watchMessageAttachmentsSaga(
+export function* watchMessagesSaga(
+  backendClient: BackendClient,
   bearerToken: SessionToken
 ): SagaIterator {
+  yield* takeLatest(
+    getType(loadNextPageMessages.request),
+    handleLoadNextPageMessages,
+    backendClient.getMessages
+  );
+
+  yield* takeLatest(
+    getType(loadPreviousPageMessages.request),
+    handleLoadPreviousPageMessages,
+    backendClient.getMessages
+  );
+
+  yield* takeLatest(
+    getType(loadPreviousPageMessages.request),
+    handleLoadPreviousPageMessages,
+    backendClient.getMessages
+  );
+
+  yield* takeLatest(
+    getType(reloadAllMessages.request),
+    handleReloadAllMessages,
+    backendClient.getMessages
+  );
+
+  yield* takeEvery(
+    loadMessageById.request,
+    handleLoadMessageById,
+    backendClient.getMessage
+  );
+
+  yield* takeLatest(
+    loadMessageDetails.request,
+    handleLoadMessageDetails,
+    backendClient.getMessage
+  );
+
+  yield* takeLatest(
+    getMessagePrecondition.request,
+    handleMessagePrecondition,
+    backendClient.getThirdPartyMessagePrecondition
+  );
+
+  yield* takeLatest(
+    loadThirdPartyMessage.request,
+    handleThirdPartyMessage,
+    backendClient.getThirdPartyMessage
+  );
+
+  yield* takeEvery(
+    upsertMessageStatusAttributes.request,
+    handleUpsertMessageStatusAttribues,
+    backendClient.upsertMessageStatusAttributes
+  );
+
+  yield* fork(watchLoadMessageData);
+
+  yield* takeLatest(
+    getType(getMessageDataAction.request),
+    handleLoadMessageData
+  );
+
+  yield* takeLatest(
+    migrateToPaginatedMessages.request,
+    handleMigrateToPagination,
+    backendClient.upsertMessageStatusAttributes
+  );
+
   // handle the request for a new downloadAttachment
   yield* takeLatest(
     downloadAttachment.request,
@@ -38,28 +121,16 @@ export function* watchMessageAttachmentsSaga(
   );
 
   // handle the request for removing a downloaded attachment
-  yield* takeEvery(
-    removeCachedAttachment,
-    function* (action: ActionType<typeof removeCachedAttachment>) {
-      yield* call(handleClearAttachment, action);
-    }
-  );
+  yield* takeEvery(removeCachedAttachment, handleClearAttachment);
 
   // handle the request for clearing user profile cache
-  yield* takeEvery(clearCache, function* () {
-    yield* call(handleClearAllAttachments);
-  });
+  yield* takeEvery(clearCache, handleClearAllAttachments);
 
   // clear cache when user explicitly logs out
-  yield* takeEvery(
-    logoutSuccess,
-    function* (_: ActionType<typeof logoutSuccess>) {
-      yield* call(handleClearAllAttachments);
-    }
-  );
+  yield* takeEvery(logoutSuccess, handleClearAllAttachments);
 }
 
-export function* watchLoadMessageData() {
+function* watchLoadMessageData() {
   yield* takeLatest(
     getType(getMessageDataAction.request),
     handleLoadMessageData
