--- ts/features/messages/saga/watchReloadAllMessages.ts
@@ -1,78 +1,0 @@
-import { call, put, takeLatest } from "typed-redux-saga/macro";
-import { ActionType, getType } from "typesafe-actions";
-import { BackendClient } from "../../../api/backend";
-import { reloadAllMessages as reloadAllMessagesAction } from "../store/actions";
-import { ReduxSagaEffect, SagaCallReturnType } from "../../../types/utils";
-import { toUIMessage } from "../store/reducers/transformers";
-import { PaginatedPublicMessagesCollection } from "../../../../definitions/backend/PaginatedPublicMessagesCollection";
-import { isTestEnv } from "../../../utils/environment";
-import { getError } from "../../../utils/errors";
-import { withRefreshApiCall } from "../../fastLogin/saga/utils";
-import { errorToReason, unknownToReason } from "../utils";
-import { trackReloadAllMessagesFailure } from "../analytics";
-import { handleResponse } from "../utils/responseHandling";
-
-type LocalActionType = ActionType<(typeof reloadAllMessagesAction)["request"]>;
-type LocalBeClient = ReturnType<typeof BackendClient>["getMessages"];
-
-export default function* watcher(
-  getMessages: LocalBeClient
-): Generator<ReduxSagaEffect, void, SagaCallReturnType<typeof getMessages>> {
-  yield* takeLatest(
-    getType(reloadAllMessagesAction.request),
-    tryReloadAllMessages(getMessages)
-  );
-}
-
-function tryReloadAllMessages(getMessages: LocalBeClient) {
-  return function* gen(
-    action: LocalActionType
-  ): Generator<ReduxSagaEffect, void, SagaCallReturnType<typeof getMessages>> {
-    const { filter, pageSize } = action.payload;
-    try {
-      const response: SagaCallReturnType<typeof getMessages> = (yield* call(
-        withRefreshApiCall,
-        getMessages({
-          enrich_result_data: true,
-          page_size: pageSize,
-          archived: filter.getArchived
-        }),
-        action
-      )) as unknown as SagaCallReturnType<typeof getMessages>;
-      const nextAction = handleResponse<PaginatedPublicMessagesCollection>(
-        response,
-        ({ items, next, prev }: PaginatedPublicMessagesCollection) =>
-          reloadAllMessagesAction.success({
-            messages: items.map(toUIMessage),
-            pagination: { previous: prev, next },
-            filter
-          }),
-        error => {
-          const reason = errorToReason(error);
-          trackReloadAllMessagesFailure(reason);
-          return reloadAllMessagesAction.failure({
-            error: getError(error),
-            filter
-          });
-        }
-      );
-
-      if (nextAction) {
-        yield* put(nextAction);
-      }
-    } catch (error) {
-      const reason = unknownToReason(error);
-      trackReloadAllMessagesFailure(reason);
-      yield* put(
-        reloadAllMessagesAction.failure({
-          error: getError(error),
-          filter
-        })
-      );
-    }
-  };
-}
-
-export const testTryLoadPreviousPageMessages = isTestEnv
-  ? tryReloadAllMessages
-  : undefined;
